+++ src/atoms/checkbox/buildTheme.ts
@@ -1,0 +1,61 @@
+import { RecursivePartial } from '@kibalabs/core';
+
+import { IBoxTheme, IColorGuide, IDimensionGuide, ITextTheme } from '../../particles';
+import { mergeTheme, ThemeMap } from '../../util';
+import { ICheckboxTheme } from './theme';
+
+export const buildCheckboxThemes = (colors: IColorGuide, dimensions: IDimensionGuide, textThemes: ThemeMap<ITextTheme>, boxThemes: ThemeMap<IBoxTheme>, base?: RecursivePartial<Record<string, ICheckboxTheme>>): ThemeMap<ICheckboxTheme> => {
+  const defaultCheckboxTheme = mergeTheme<ICheckboxTheme>({
+    normal: {
+      default: {
+        checkBackground: mergeTheme(boxThemes.default, boxThemes.focusable, {
+          padding: `${dimensions.paddingNarrow3} ${dimensions.paddingNarrow3}`,
+          'border-width': '2px',
+          'border-color': '$colors.brandPrimary',
+          'border-radius': '0.2em',
+        }),
+        text: mergeTheme(textThemes.default, {
+        }),
+        checkColor: '$colors.textOnBrand',
+      },
+      hover: {
+        checkBackground: {
+          'background-color': '$colors.brandPrimaryClear80',
+        },
+        text: {
+          color: '$colors.brandPrimary',
+        },
+      },
+      press: {
+        checkBackground: {
+          'background-color': '$colors.brandPrimaryClear50',
+        },
+      },
+      focus: {
+        checkBackground: boxThemes.focussed,
+      },
+    },
+    checked: {
+      default: {
+        checkBackground: {
+          'background-color': '$colors.brandPrimary',
+        },
+      },
+    },
+    disabled: {
+      default: {
+        checkBackground: {
+          'background-color': '$colors.disabled',
+        },
+        text: {
+          color: '$colors.disabledText',
+        },
+      },
+    },
+  }, base?.default);
+
+  return {
+    ...base,
+    default: defaultCheckboxTheme,
+  };
+};